added some views with controllers

+++ database/migrations/2020_09_02_173430_create_orders_table.php
@@ -9,7 +9,7 @@ class CreateOrdersTable extends Migration {
 	{
 		Schema::create('orders', function(Blueprint $table) {
 			$table->increments('id');
-			$table->integer('user_id')->unsigned();
+			$table->bigInteger('user_id')->unsigned();
 			$table->integer('product_id')->unsigned();
 			$table->string('delivery');
 			$table->timestamps();
